Dirty fix in DCS-BIOS when SelectedDCSBIOSInput is null
