+++ Source/DCS-BIOS/DCS-BIOS.csproj
@@ -135,6 +135,7 @@
   <ItemGroup>
     <Compile Include="DBCommon.cs" />
     <Compile Include="DCSBIOS.cs" />
+    <Compile Include="DCSBIOSInputObject.cs" />
     <Compile Include="Json\DCSBIOSControl.cs" />
     <Compile Include="Json\DCSBIOSControlInput.cs" />
     <Compile Include="DCSBIOSControlLocator.cs" />
